+++ src/main/java/it/unibo/heavypocket/mvc/model/core/ExpenseImpl.java
@@ -4,12 +4,13 @@ import java.time.LocalDate;
 import java.util.Collections;
 import java.util.Objects;
 import java.util.Set;
+import java.math.BigDecimal;
 
 public class ExpenseImpl extends AbstractTransaction implements Expense {
 
     private final Set<Tag> tags;
 
-    public ExpenseImpl(final String description, final double amount, final LocalDate date, final Set<? extends Tag> tags) {
+    public ExpenseImpl(final String description, final BigDecimal amount, final LocalDate date, final Set<? extends Tag> tags) {
         super(description, amount, date);
         this.tags = Set.copyOf(tags);
     }
